Make querying possible with jj-abrams branch
1) Remove dashboard-druid-client 'cause I think it's trying to abstract a bit too much.  As this starts being more dynamic, we will need to provide more, different parameters than what it was trying to allow
2) Add things to druid-client to make up for what was removed from dashboard-druid-client
3) Various other small adjustments

With these changes, it is possible to get data to query and show, however, there appear to still be things to be done on the visualization side of things.  Also, this branch has lost support for filtering (including removing of filters) as well as setting which metric the topN lists should be done on.  That will have to be fixed before this can be merged into master.